Week6-Lab-Submission-ForNCI
Take a copy of your existing CV site and modify it to use Bootstrap as its grid system.

You may also use any of the Bootstrap components that you like, but you will be graded on your responsive design using Bootstrap.

When you have completed your work, commit it to GitHub and submit a Word/Text document with a link to the repository.